NXDRIVE-2063: Move tests logic to tox
Part of the tests are now managed by `tox`, to run all:

    tox

To run a specific test environment, check the current list using:

    tox -l

And then run the desired environment:

    TOXENV=unit tox

You can select what test to run using the `-k` selector
(see "Using -k expr to select tests based on their name" at
http://doc.pytest.org/en/latest/example/markers.html for details):

    TOXENV=unit tox -- -k url
    TOXENV=unit tox -- -k test_utils
    TOXENV=unit tox -- -k test_utils.py

Note: I also automated the retrieval of the `SENTRY_ENV` value
based on the current branch name. It is no more needed to set
that envar manually.
