+++ docs/changes/4.4.2.md
@@ -58,6 +58,7 @@ Release date: `20xx-xx-xx`
 - [NXDRIVE-1994](https://jira.nuxeo.com/browse/NXDRIVE-1994): [Windows] Skip JUnit report when running a specific test
 - [NXDRIVE-1996](https://jira.nuxeo.com/browse/NXDRIVE-1996): Add a script to check translations files
 - [NXDRIVE-2013](https://jira.nuxeo.com/browse/NXDRIVE-2013): Create a script to convert NCO CSV log files to real log files
+- [NXDRIVE-2063](https://jira.nuxeo.com/browse/NXDRIVE-2063): Move tests logic to tox
 
 ## Docs
 
@@ -99,6 +100,7 @@ Release date: `20xx-xx-xx`
 - Upgraded `urllib3` from 1.25.6 to 1.25.8
 - Upgraded `watchdog` from commit-6793588 to 0.10.2
 - Upgraded `xattr` from 0.9.6 to 0.9.7
+- Testing: The `SENTRY_ENV` envar is now automatically retrieved based on the branch name
 
 ## Technical Changes
