+++ docs/deployment.md
@@ -24,7 +24,6 @@ Possible `ARG`:
     --install-python: install only Python
     --install-release: install a complete environment for a release (without tests requirements)
     --start: start Nuxeo Drive
-    --check-translations: check the translation files
     --tests: launch the tests suite
 
 Executing the script without argument will setup/update the isolated environment.
@@ -86,15 +85,7 @@ If you get an error message complaining about the lack of signature for this scr
 ```
 - `SKIP` is used to tweak tests checks:
 ```
-    - SKIP=flake8 to skip code style
-    - SKIP=spell to skip grammar check
-    - SKIP=translations to skip translation files check
-    - SKIP=mypy to skip type annotations
-    - SKIP=cleanup to skip dead code checks
     - SKIP=rerun to not rerun failed test(s)
-    - SKIP=integration to not run integration tests on Windows
-    - SKIP=all to skip all above (equivalent to flake8,mypy,rerun,translations,spell,integration)
-    - SKIP=tests tu run only code checks
 ```
 
 #### MacOS Specific
