--- requirements-tests.txt
@@ -1,14 +1,0 @@
-codespell==1.16.0
-faker==4.0.1
-flake8==3.7.9
-junitparser==1.4.1
-mypy==0.761
-pre-commit==2.1.1
-pytest==5.3.5
-pytest-cov==2.8.1
-pytest-benchmark==3.2.3
-pytest-timeout==1.3.4
-pytest-xdist==1.31.0
-pywinauto==0.6.8; sys_platform == 'win32'
-selenium==3.141.0; sys_platform == 'win32'
-vulture==1.3
