+++ setup.cfg
@@ -35,5 +35,3 @@ addopts =
     --durations=20
     # Print a full stacktrace of all threads after 20 seconds
     # --faulthandler-timeout=195
-    # Disable the benchmark plugin, it will be used manually only
-    --benchmark-disable
