+++ tools/deps/requirements-cleanup.txt
@@ -1,0 +1,2 @@
+vulture==1.3 \
+    --hash=sha256:54f013779e071c7e1657e3a0b567f7400e046a8beb08585179ffbe762661fb5e
