+++ tools/deps/requirements-spell.txt
@@ -1,0 +1,2 @@
+codespell==1.16.0 \
+    --hash=sha256:a81780122f955002d032a9a9c55e2305c6f252a530d8682ed5c3d0580f93d9b8
