+++ tools/deps/requirements-types.txt
@@ -1,0 +1,11 @@
+mypy==0.761 \
+    --hash=sha256:7eadc91af8270455e0d73565b8964da1642fe226665dd5c9560067cd64d56749
+mypy-extensions==0.4.3 \
+    --hash=sha256:090fedd75945a69ae91ce1303b5824f428daf5a028d2f6ab8a299250a846f15d
+    # via mypy
+typed-ast==1.4.1 \
+    --hash=sha256:bcd3b13b56ea479b3650b82cabd6b5343a625b0ced5429e4ccad28a8973f301b
+    # via mypy
+typing-extensions==3.7.4.1 \
+    --hash=sha256:cf8b63fedea4d89bab840ecbb93e75578af28f76f66c35889bd7065f5af88575
+    # via mypy
